[BACKEND-14] refactor(core): configure_logger и get_logger добавлены в src/core/init.py, обновлены импорты
