+++ src/api/lifespan.py
@@ -2,8 +2,7 @@ from contextlib import asynccontextmanager
 
 from fastapi import FastAPI
 
-from src.core.logger.config import configure_logger
-from src.core.logger.logger import get_logger
+from src.core.logger import configure_logger, get_logger
 
 # from src.core.db import init_db_connection
 # from src.repository.s3.s3 import init_s3_client
