+++ src/core/logger/__init__.py
@@ -1,4 +1,7 @@
 from .config import configure_logger
 from .logger import get_logger
 
-__all__ = ["get_logger", "configure_logger"]
+__all__ = [
+    "configure_logger",
+    "get_logger"
+]
